Split Fees module out of Cardano.Api.Typed
